Replace ST style tests with more FV style tests
Rather than trying to verify that packets can actually flow (which
requires felix to run etc...) these tests just verify that the write
information is written to the datastore and the right network devices
are created.

The more extensive ST style tests will be moved to calico-containers
where they belong.